lab-04/requirements.txt
markdown-it-py
mdurl
numpy
pillow
Pygments
rich
shapely
stopit
types-pillow
setuptools
